# openapi.yaml
openapi: 3.0.3
info:
  title: Airport Transfer Marketplace API
  version: 1.1.0
  description: |
    Independent Airport Transfer Marketplace API.
    Covers authentication, suppliers, airports/zones/routes, pricing, bookings, rides, and payments.

    ## Role-Based Access Control

    - `GET /public/*`, `POST /public/*`, `/webhooks/*`: **no authentication**
    - `/auth/*`: no authentication
    - `/supplier/*`: requires JWT with roles: `SUPPLIER_OWNER`, `DISPATCHER`, or `DRIVER`
    - `/admin/*`: requires JWT with role: `ADMIN`

    ## Currency Handling

    - Marketplace operates in **EUR** as base currency
    - Suppliers can set prices in their local currency (TRY, USD, GBP, etc.)
    - Currency conversion happens at booking time using real-time rates
    - Supplier payouts are made in their configured payout currency

    ## Cancellation Policy

    - Free cancellation up to **24 hours** before pickup
    - **50% refund** for cancellations **12–24 hours** before pickup
    - **No refund** for cancellations less than **12 hours** before pickup
    - Supplier cancellations result in full refund + possible compensation
servers:
  - url: https://api.airporttransferportal.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotificationRecipientType:
      type: string
      enum: [CUSTOMER, DRIVER, SUPPLIER, DISPATCHER, AGENCY]
    NotificationType:
      type: string
      enum: [BOOKING_CONFIRMED, DRIVER_ASSIGNED, DRIVER_EN_ROUTE, DRIVER_ARRIVED, TRIP_STARTED, TRIP_COMPLETED, FLIGHT_DELAYED, PICKUP_TIME_CHANGED, BOOKING_CANCELLED, REVIEW_REQUEST, PAYMENT_RECEIVED, PAYOUT_SENT, DOCUMENT_EXPIRING, GENERAL]
    NotificationChannel:
      type: string
      enum: [EMAIL, SMS, PUSH, WHATSAPP, IN_APP]
    NotificationStatus:
      type: string
      enum: [PENDING, SENT, DELIVERED, FAILED, READ]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        recipientType:
          $ref: '#/components/schemas/NotificationRecipientType'
        recipientId:
          type: integer
          nullable: true
        recipientEmail:
          type: string
          nullable: true
        recipientPhone:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        message:
          type: string
        bookingId:
          type: integer
          nullable: true
        bookingCode:
          type: string
          nullable: true
        rideId:
          type: integer
          nullable: true
        channel:
          $ref: '#/components/schemas/NotificationChannel'
        status:
          $ref: '#/components/schemas/NotificationStatus'
        sentAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /dispatch/notifications:
    get:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Get notifications
      description: Returns notifications for dispatch operations
      parameters:
        - name: recipientType
          in: query
          schema:
            $ref: '#/components/schemas/NotificationRecipientType'
        - name: recipientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: channel
          in: query
          schema:
            $ref: '#/components/schemas/NotificationChannel'
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
    post:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Create notification
      description: Create a new notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipientType, type, title, message, channel]
              properties:
                recipientType:
                  $ref: '#/components/schemas/NotificationRecipientType'
                recipientId:
                  type: integer
                recipientEmail:
                  type: string
                recipientPhone:
                  type: string
                type:
                  $ref: '#/components/schemas/NotificationType'
                title:
                  type: string
                message:
                  type: string
                bookingId:
                  type: integer
                rideId:
                  type: integer
                channel:
                  $ref: '#/components/schemas/NotificationChannel'
                metadata:
                  type: object
      responses:
        '200':
          description: Notification created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notificationId:
                    type: integer
    put:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Update notification status
      description: Update notification status or mark as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                ids:
                  type: array
                  items:
                    type: integer
                status:
                  $ref: '#/components/schemas/NotificationStatus'
                markAsRead:
                  type: boolean
      responses:
        '200':
          description: Notification updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean